+++ docs/sample/example_aws.yaml
