+++ docs/sample/example_gcloud.yaml
