+++ docs/sample/example_linode.yaml
