+++ docs/sample/example_oci.yaml
