+++ internal/conf/globals.go
