+++ internal/distlog/type_logbot.go
@@ -1,4 +1,4 @@
-package tuipack
+package distlog
 
 import (
 	"fmt"
