+++ internal/distlog/type_logmsg.go
@@ -1,4 +1,4 @@
-package tuipack
+package distlog
 
 import (
 	"fmt"
