+++ internal/provider/type_job.go
@@ -2,7 +2,7 @@ package provider
 
 import (
 	"github.com/orme292/objectify"
-	"github.com/orme292/s3packer/conf"
+	"github.com/orme292/s3packer/internal/conf"
 )
 
 type Job struct {
