+++ internal/provider/type_pathmodemap.go
