+++ internal/providers/aws/globals.go
