+++ internal/providers/gcloud/globals.go
