+++ internal/providers/gcloud/object.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/orme292/s3packer/s3packs/provider"
+	"github.com/orme292/s3packer/internal/provider"
 )
 
 type CloudObject struct {
