+++ internal/providers/oracle/globals.go
